requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
pdfminer.six==20231228
python-docx==1.1.2
python-dotenv==1.0.1
google-genai==0.3.0